Moved ValidationResponse to its own class file.
